Revert inline-frame work
Because I get this error in Firefox:
```
Blocked by Content Security Policy

An error occurred during a connection to github.com.

Firefox prevented this page from loading in this way because the page has a content security policy that disallows it.
```

(Even if using github instead of gitlab.)

In Brave(Chrome) it doesn't even load at, fails silently with no
error, but I guess it's because of the same reason.